Moved add_file() and get_template_string() to utils/plugin_utils.py.
